requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.6.4
deepdiff==6.7.1
Deprecated==1.2.14
distro==1.9.0
emoji==2.11.0
fastapi==0.110.1
filelock==3.13.3
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-experimental==0.0.55
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.33
lxml==5.2.1
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.14.3
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pillow==10.3.0
poppler-utils==0.1.0
posthog==3.5.0
protobuf==4.25.3
psycopg2==2.9.9
pulsar-client==3.4.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMySQL==1.1.0
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.7.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
types-requests==2.31.0.20240403
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unstructured==0.11.8
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
